+++ CHANGELOG.md
@@ -37,6 +37,7 @@ Changelog
 * Added restriction for *getPackagesForUid*
 * Added process name parameter to *queryContentProviders*
 * Added restriction for heart rate sensor (Android "L") ([issue](/../../issues/1757))
+* Added restriction *SIP.isSipWifiOnly*
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)